move os_realestate models to stopo_master
