+++ production.ini.in
@@ -29,7 +29,6 @@ sqlalchemy.kogis.url = postgresql://${dbhost}:${dbport}/kogis_${dbstaging}
 sqlalchemy.stopo.url = postgresql://${dbhost}:${dbport}/stopo_${dbstaging}
 sqlalchemy.uvek.url = postgresql://${dbhost}:${dbport}/uvek_${dbstaging}
 sqlalchemy.uvek_solarkataster.url = postgresql://${dbhost}:${dbport}/uvek_solarkataster_${dbstaging}
-sqlalchemy.mogis.url = postgresql://${dbhost}:${dbport}/mogis_${dbstaging}
 sqlalchemy.vbs.url = postgresql://${dbhost}:${dbport}/vbs_${dbstaging}
 sqlalchemy.zeitreihen.url = postgresql://${dbhost}:${dbport}/zeitreihen_${dbstaging}
 sqlalchemy.lubis.url = postgresql://${dbhost}:${dbport}/lubis_${dbstaging}
